requirements.txt
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
download==0.3.5
feedparser==6.0.10
Flask==2.2.2
fuzzywuzzy==0.18.0
hurry.filesize==0.9
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==22.2.2
python-qbittorrent==0.4.2
python3-wget==0.0.2b1
PyYAML==6.0
rarfile==4.0
requests==2.28.1
setuptools==65.3.0
sgmllib3k==1.0.0
six==1.16.0
tqdm==4.64.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1

